feat-chore: openai_images package to manage code
