+++ autobots/action/action_type/action_text2img/action_text2img_dalle_openai_v2.py
@@ -6,7 +6,7 @@ from autobots.action.action_type.abc.IAction import IAction, ActionOutputType, A
 from autobots.action.action.action_doc_model import ActionCreate
 from autobots.action.action_type.action_types import ActionType
 from autobots.action.action.common_action_models import TextObj
-from autobots.conn.openai.image_model import ImageReq
+from autobots.conn.openai.openai_images.image_model import ImageReq
 from autobots.conn.openai.openai_client import get_openai
 
 
@@ -37,5 +37,5 @@ class ActionGenImageDalleOpenAiV2(IAction[ImageReq, TextObj, ImagesResponse]):
 
     async def run_action(self, action_input: TextObj) -> ImagesResponse:
         self.action_config.prompt = f"{self.action_config.prompt}\n{action_input.text}"
-        images = await get_openai().create_image(self.action_config)
+        images = await get_openai().openai_images.create_image(self.action_config)
         return images
