+++ autobots/conn/openai/openai_images/image_model.py
